+++ Geome3Dash/src/game/planet/G3DPlanetLayer.cpp
@@ -392,7 +392,7 @@ namespace g3d
                 continue;
             }
 
-            const auto percentage = PlanetStateManager::getInstance()->getProgressByLevelID(i - 1)->normal;
+            const auto percentage = PlanetStateManager::getInstance()->getProgressByLevelID(static_cast<int>(i) - 1)->normal;
 
             cloudModel->meshes[realMeshId]->setVisible(percentage == 100 ? 0 : 1);
         }
@@ -562,7 +562,7 @@ namespace g3d
             shaderProgram->setVec3("Ks", isCustomKs ? customKs : Ks);
             shaderProgram->setFloat("shininess", 32);
             shaderProgram->setFloat("d", (isCustomD ? customD : d) * opacityModifier);
-            glDrawElements(GL_TRIANGLES, indices.size(), GL_UNSIGNED_INT, NULL);
+            glDrawElements(GL_TRIANGLES, static_cast<int>(indices.size()), GL_UNSIGNED_INT, NULL);
             glBindVertexArray(0);
             glDisable(GL_DEPTH_TEST);
             glDisable(GL_BLEND);
